refactor&remove: files and folders
