+++ fronteend/src/components/CardStatusDaLixeira.jsx
