+++ fronteend/src/components/Navbar.jsx
@@ -13,7 +13,7 @@ import SearchIcon from '@mui/icons-material/Search';
 import AccountCircle from '@mui/icons-material/AccountCircle'; 
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import MoreIcon from '@mui/icons-material/MoreVert';
-import MenuLateral from '../MenuLateral';
+import MenuLateral from './MenuLateral';
 
 
 const Search = styled('div')(({ theme }) => ({
