+++ fronteend/src/pages/Dashboard/StatusDaLixeira.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import MiniDrawer from '../../components/Drawer'
+import MiniDrawer from '../../components/MiniDrawer'
 //import Sidebar from '../../components/Sidebar'
 
 export const StatusDaLixeira = () => {
